refactor: split manifest_validator.go into focused validator modules
Break down the 2753-line manifest_validator.go into six focused files,
each under 600 lines, organized by validation domain:

- manifest_validator.go (555): Core types, constructor, orchestrator, shared helpers
- cel_validator.go (541): CEL expression validation, AsyncAPI schema loading
- starlark_validator.go (396): Starlark compilation, predeclared setup, handler enrichment
- crossref_validator.go (567): Cross-reference and duplicate detection
- domain_validator.go (494): Payment rails, party types, mappings, triggers, API validation
- lifecycle_validator.go (248): Immutability checks, destructive change detection, orphan detection

Pure refactor — no behavioral changes. All functions remain in package validator
with identical signatures.
